requirements.txt
# Core analysis
scanpy>=1.9.0
anndata>=0.10.0
numpy>=1.23
pandas>=2.0
scipy>=1.10
matplotlib>=3.7

# Integration models
scvi-tools>=1.1.0
torch>=2.0

# Scripts / notebooks
rich>=13.0
jupyter>=1.0
ipywidgets>=8.0

# Optional: doublet scoring in preprocessing.ipynb
scrublet>=0.2

# Optional: R-backed steps in preprocessing.ipynb (CellxGene / some QC)
# rpy2>=3.5

# Benchmarking (step 7)
scib-metrics>=0.5.0

# Optional: GPU k-NN for scANVI_benchmark.py (RAPIDS cuML)
# cuml-cu12  # match to your CUDA stack via conda or NVIDIA docs
